src/test/java/guru/springframework/jdbc/dao/AuthorDaoHibernateTest.java
package guru.springframework.jdbc.dao;

import guru.springframework.jdbc.domain.Author;
import jakarta.persistence.EntityManagerFactory;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
import org.springframework.data.domain.PageRequest;
import org.springframework.data.domain.Pageable;
import org.springframework.data.domain.Sort;
import org.springframework.test.context.ActiveProfiles;

import java.util.List;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;

@ActiveProfiles("local")
@DataJpaTest
@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
public class AuthorDaoHibernateTest {
    @Autowired
    EntityManagerFactory emf;

    AuthorDao authorDao;

    @BeforeEach
    void setup() {
        authorDao = new AuthorDaoHibernate(emf);
    }

    @Test
    void findAllAuthorsSortByLastName() {
        Pageable pageable = PageRequest.of(0, 10, Sort.by(Sort.Order.desc("firstName")));
        List<Author> books = authorDao.findAllAuthorsByLastName("Smith", pageable);

        assertThat(books).isNotNull();
        assertThat(books.size()).isEqualTo(10);
    }
}
